passing the UserId to the client side

+++ mean-stack/backend/routes/posts.js
@@ -70,8 +70,12 @@ router.put('/:id', checkAuth, multer({storage:storage}).single("image"), (req,re
         content:req.body.content,
         imagePath:imagePath
     });
-    Post.updateOne({_id: req.params.id}, post).then(result=>{
-        res.status(200).json({message:'Update successful!', post:{...post, id:req.params.id}});
+    Post.updateOne({_id: req.params.id, creator:req.userData.userId}, post).then(result=>{
+        if(result.nModified > 0){
+            res.status(200).json({message:'Update successful!', post:{...post, id:req.params.id}});
+        }else{
+            res.status(401).json({message:'Not Authorized!'});
+        }
     });
 })
 
@@ -116,8 +120,12 @@ router.delete('/:id', checkAuth, (req,res,next) => {
     // });
 
     //to delete post from database
-    Post.deleteOne({_id:req.params.id}).then(result=>{
-        res.status(200).json({message:'The Post deleted!'})
+    Post.deleteOne({_id:req.params.id, creator:req.userData.userId}).then(result=>{
+        if(result.n > 0){
+            res.status(200).json({message:'The Post deleted!'})
+        }else{
+            res.status(401).json({message:'Not Authorized!'});
+        }
     })
 })
